Independent_Projects/AWS/Indeed_Pipeline/artifacts/docker_image/python-docx-environment/requirements.txt
pandas
boto3
python-docx
fsspec
s3fs